+++ docs/changelog.rst
@@ -20,6 +20,8 @@ Other changes:
 
 - Added VAT identifcation number validator for all EU locales.
 
+- Added German translation for BIC (Bank Identifier Code) error messages.
+
 2.0   (2017-12-30)
 ------------------