+++ docs/changelog.rst
@@ -11,6 +11,7 @@ New flavors:
 New fields for existing flavors:
 
 - `NLLicensePlateField` in NL flavor.
+- `GRSocialSecurityNumberField` (AMKA) in GR flavore
 
 Modifications to existing flavors: