+++ docs/changelog.rst
@@ -21,6 +21,7 @@ Other changes:
 - Added VAT identifcation number validator for all EU locales.
 - Fix EAN validation when intermediate checksum is 10
   (`gh-331 <https://github.com/django/django-localflavor/issues/331>`_).
+- Confirmed support for Django 2.1.
 
 
 2.0   (2017-12-30)